BackEnd/requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
sqlparse==0.5.0
tzdata==2024.1
psycopg2
drf-spectacular==0.27.2
Pillow==10.3.0
factory_boy